Phisher2025-GiT ready/requirements.txt
# Core ML/DL
tensorflow==2.20.0
numpy==2.2.6
pandas==2.3.3
scikit-learn==1.7.2

# PyTorch libraries (found in venv)
torch==2.9.1
pytorch-lightning==2.5.6
torchmetrics==1.8.2
transformers==4.57.1

# API & Serving
fastapi==0.121.1
uvicorn==0.38.0
requests==2.32.5
python-json-logger==2.0.7

# Config & Utils
PyYAML==6.0.3
python-dotenv==1.2.1

# Testing
pytest==9.0.1
pytest-cov==7.0.0
